requirements.txt
streamlit-lottie
matplotlib
streamlit
sentence-transformers
mistralai
pandas
numpy
scikit-learn
xgboost
optuna
Pillow
plotly